Add RKNN support

+++ ultralytics/data/augment.py
@@ -501,8 +501,8 @@ class RandomPerspective:
         # Create new boxes
         x = xy[:, [0, 2, 4, 6]]
         y = xy[:, [1, 3, 5, 7]]
-        return np.concatenate((x.min(1), y.min(1), x.max(1), y.max(1)), dtype=bboxes.dtype).reshape(4, n).T
-
+        # return np.concatenate((x.min(1), y.min(1), x.max(1), y.max(1)), dtype=bboxes.dtype).reshape(4, n).T
+        return np.concatenate((x.min(1), y.min(1), x.max(1), y.max(1))).reshape(4, n).T
     def apply_segments(self, segments, M):
         """
         Apply affine to segments and generate new bboxes from segments.
